+++ spec/helpers/claims_helper_spec.rb
@@ -44,12 +44,12 @@ describe ClaimsHelper do
 
       it "returns an array of identity-related questions and answers for displaying to the user for review" do
         expected_answers = [
-          [I18n.t("questions.name"), "Jo Bloggs", "name"],
+          [I18n.t("questions.name"), "Jo Bloggs", "personal-details"],
           [I18n.t("questions.address.generic.title"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
-          [I18n.t("questions.date_of_birth"), "10 January 1980", "date-of-birth"],
+          [I18n.t("questions.date_of_birth"), "10 January 1980", "personal-details"],
           [I18n.t("questions.payroll_gender"), "Don’t know", "gender"],
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
-          [I18n.t("questions.national_insurance_number"), "QQ123456C", "national-insurance-number"],
+          [I18n.t("questions.national_insurance_number"), "QQ123456C", "personal-details"],
           [I18n.t("questions.email_address"), "test@email.com", "email-address"]
         ]
 
@@ -61,7 +61,7 @@ describe ClaimsHelper do
 
         expected_answers = [
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
-          [I18n.t("questions.national_insurance_number"), "QQ123456C", "national-insurance-number"],
+          [I18n.t("questions.national_insurance_number"), "QQ123456C", "personal-details"],
           [I18n.t("questions.email_address"), "test@email.com", "email-address"]
         ]
 
@@ -72,12 +72,12 @@ describe ClaimsHelper do
         claim.date_of_birth = nil
 
         expected_answers = [
-          [I18n.t("questions.name"), "Jo Bloggs", "name"],
+          [I18n.t("questions.name"), "Jo Bloggs", "personal-details"],
           [I18n.t("questions.address.generic.title"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
-          [I18n.t("questions.date_of_birth"), nil, "date-of-birth"],
+          [I18n.t("questions.date_of_birth"), nil, "personal-details"],
           [I18n.t("questions.payroll_gender"), "Don’t know", "gender"],
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
-          [I18n.t("questions.national_insurance_number"), "QQ123456C", "national-insurance-number"],
+          [I18n.t("questions.national_insurance_number"), "QQ123456C", "personal-details"],
           [I18n.t("questions.email_address"), "test@email.com", "email-address"]
         ]